Abstract the delete functions and style classes to be for general use.
