+++ src/main/webapp/stylesheets/listing-preview.css
@@ -1,41 +1,3 @@
-.cancel-button {
-  background-color: #8EEEDE;
-  margin-left: .5rem;
-}
-
-.delete-button {
-  background-color: #F20202;
-  margin-right: .5rem;
-}
-
-.delete-button-container {
-  display: flex;
-}
-
-.delete-container {
-  align-items: center;
-  border-radius: 2rem;
-  background-color: white;
-  display: inline-flex;
-  flex-direction: column;
-  padding: 1rem;
-}
-
-.delete-modal {
-  cursor: default;
-}
-
-.delete-modal-button {
-  border-radius: 1rem;
-  cursor: pointer;
-  padding: .3rem;
-}
-
-.error-listing-name {
-  color: #42857A;
-  font-weight: 700;
-}
-
 .listing {
   display: flex;
   flex-flow: row nowrap;
@@ -80,17 +42,6 @@
   align-items: center;
   display: flex;
 }
-
-.material-icons.delete-icon:hover {
-  color: red;
-}
-
-.material-icons.modal-delete-icon {
-  color: red;
-  font-size: 3rem;
-}
-
-.owner-
 
 .preview-info {
   align-items: center;
